timetable_scheduler/requirements.txt
streamlit>=1.32.0
mysql-connector-python>=8.3.0
pandas>=2.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
bcrypt>=4.0.0
streamlit-authenticator>=0.3.1
Pillow>=10.0.0